Improved figcaptions
Changed the min-width of figure elements from 95% to 400px.

The min-width is there so that the figcaption has some width
even if the figure content does not load. In normal situations
with 95% width the figcaption content will be wider than the
figure when the figure content is narrow.

As we use 'display: table' and 'display: table-caption' with the
figure and figcaption more narrow min-width allows the figcaption
to be the same width as the figure content (image, video etc.)
as long as the content is wider than the min-width (in this
case 400px).