Added a breakpoint for the min-width of the figure element
The min-width of 400px for figure elements is now activated
only if the width of the viewport is at least 500px.

This allows the figures to resize to smaller widths on narrow
viewports.

The min-width is there for the case that the figure content
does not load. In that case the figcaption has some more width
thanks to the min-width.